Move CLI utils to `dwave.cloud.utils.cli`
